Create question_6.c
question 6 - 
Write a C/Java program which prints the area and volume of any one of the given shapes
given below. Accept the choice of the shape, appropriate inputs from the user, calculate
and display the area and the volume of the same. Repeat this with different shapes till
the user wishes to stop.
Cylinder: Area : A=2πrh+2πr2

Volume: V=πr2
h

Cone: Area: A=πr(r+ √(h
2+r2
)) Volume: V= πr2
h/3
Sphere: Area: A= 4πr2 Volume: V= (4/3) πr3